I've implemented the user profile page and fixed the group image upload.
- I implemented the user profile page, which now shows your posts and reels in a grid.
- The profile page now correctly displays an 'Edit Profile' button for you and 'Follow'/'Message' buttons for other users.
- I fixed the group creation flow by switching the group profile image upload from Supabase Storage to Firebase Storage. This resolves the issue where you could not upload a profile image when creating a new group.
